+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres roughly to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Set `data-new-page-link` on `Link`s when the link opens a new page (so that it can be disabled in the UI once clicked)
 ### Fixed
+- a `FormButton` that is remote will briefly disable instead of disable.
 
 ## [0.6.0] - 2022-08-28
 ### Added